// RentACarProjectAngular-master/src/app/models/car.ts
export interface Car {
  carId: number;
  brandId: number;
  colorId: number;
  branchId: number;
  gearId: number;
  fuelId: number;
  classId: number;
  caseId: number;
  carPlate: string;
  carStar: number;
  modelYear: number;
  dailyPrice: number;
  description: string;
  carUsable: string;
  carLocation: string;
}